+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Improve automatic schema migration for Hive and JDBC relations
 * Improve support of CHAR(n) and VARCHAR(n) types. Those types will now be propagates to Hive
+* Fix the name of some config variables (floman. => flowman.)
+* Added new config variables `flowman.default.relation.migrationPolicy` and `flowman.default.relation.migrationStrategy`
 
 
 # Version 0.17.1 - 2021-06-18